requirements.txt
pyqrcode
pypng
fpdf
pytube
langdetect